requirements-core.txt
# Core Django Requirements - Essential packages only
Django>=4.2.16,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
Pillow>=10.1.0
drf-yasg>=1.21.7
uritemplate>=4.1.1
packaging>=23.2
